# api.yml
openapi: 3.0.3
info:
  title: Place-your-bets service
  version: 1.0.0
  description: An API intended to be consumed by the place-your-bets frontend

paths:

  /health:
    get:
      summary: Just an endpoint to see that the backend is running. Always returns 200
      operationId: health
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object

  /info:
    get:
      summary: Get all the info the frontend needs in one call
      operationId: getInfo
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: Contains the player ID. Not real authentication!
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  bitcoinPrice:
                    type: number
                    nullable: true
                    description: The current Bitcoin price, or null if unavailable
                  recentGuesses:
                    type: array
                    items:
                      type: object
                      properties:
                        priceAtSubmission:
                          type: number
                        submittedAt:
                          type: number
                        direction:
                          type: string
                          enum:
                            - "UP"
                            - "DOWN"
                        priceAtResolution:
                          type: number
                        resolvedAt:
                          type: number
                        result:
                          type: string
                          enum:
                            - "CORRECT"
                            - "INCORRECT"
                      required:
                        - priceAtSubmission
                        - submittedAt
                        - direction
                  score:
                    type: number
                required:
                  - bitcoinPrice
                  - recentGuesses
                  - score
  /submit-guess:
    post:
      summary: Submit a guess
      operationId: submitGuess
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: Contains the player ID. Not real authentication!
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                direction:
                  type: string
                  enum:
                    - "UP"
                    - "DOWN"
              required:
                - direction
      responses:
        '201':
          description: Success